requirements.txt
asgiref==3.5.2
astroid==2.11.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
httpie==3.2.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
platformdirs==2.5.2
psycopg2==2.9.3
Pygments==2.12.0
pylint==2.14.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
